+++ src/main/java/com/SOLUX_WEBFINITE_BE/webfinite_be/repository/CourseRepository.java
@@ -1,4 +1,4 @@
-package com.SOLUX_WEBFINITE_BE.webfinite_be.reposiroty;
+package com.SOLUX_WEBFINITE_BE.webfinite_be.repository;
 
 import com.SOLUX_WEBFINITE_BE.webfinite_be.domain.Course;
 import com.SOLUX_WEBFINITE_BE.webfinite_be.domain.CourseFile;
