+++ src/main/java/com/SOLUX_WEBFINITE_BE/webfinite_be/repository/PromptRepository.java
@@ -1,4 +1,4 @@
-package com.SOLUX_WEBFINITE_BE.webfinite_be.reposiroty;
+package com.SOLUX_WEBFINITE_BE.webfinite_be.repository;
 
 import com.SOLUX_WEBFINITE_BE.webfinite_be.domain.Prompt;
 import org.springframework.data.jpa.repository.JpaRepository;
