+++ src/main/java/com/SOLUX_WEBFINITE_BE/webfinite_be/service/CourseService.java
@@ -9,9 +9,9 @@ import com.SOLUX_WEBFINITE_BE.webfinite_be.exception.CourseNotFoundException;
 import com.SOLUX_WEBFINITE_BE.webfinite_be.exception.EmptyFileContentException;
 import com.SOLUX_WEBFINITE_BE.webfinite_be.exception.FileNotFoundException;
 import com.SOLUX_WEBFINITE_BE.webfinite_be.exception.UserNotFoundException;
-import com.SOLUX_WEBFINITE_BE.webfinite_be.reposiroty.CourseRepository;
-import com.SOLUX_WEBFINITE_BE.webfinite_be.reposiroty.FileRepository;
-import com.SOLUX_WEBFINITE_BE.webfinite_be.reposiroty.UserRepository;
+import com.SOLUX_WEBFINITE_BE.webfinite_be.repository.CourseRepository;
+import com.SOLUX_WEBFINITE_BE.webfinite_be.repository.FileRepository;
+import com.SOLUX_WEBFINITE_BE.webfinite_be.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
