+++ src/main/java/com/SOLUX_WEBFINITE_BE/webfinite_be/service/PlanService.java
@@ -9,10 +9,10 @@ import com.SOLUX_WEBFINITE_BE.webfinite_be.dto.SimpleResponse;
 import com.SOLUX_WEBFINITE_BE.webfinite_be.dto.gpt.GeneratePrompt;
 import com.SOLUX_WEBFINITE_BE.webfinite_be.exception.*;
 import com.SOLUX_WEBFINITE_BE.webfinite_be.exception.FileNotFoundException;
-import com.SOLUX_WEBFINITE_BE.webfinite_be.reposiroty.CourseRepository;
-import com.SOLUX_WEBFINITE_BE.webfinite_be.reposiroty.FileRepository;
-import com.SOLUX_WEBFINITE_BE.webfinite_be.reposiroty.PlanRepository;
-import com.SOLUX_WEBFINITE_BE.webfinite_be.reposiroty.PromptRepository;
+import com.SOLUX_WEBFINITE_BE.webfinite_be.repository.CourseRepository;
+import com.SOLUX_WEBFINITE_BE.webfinite_be.repository.FileRepository;
+import com.SOLUX_WEBFINITE_BE.webfinite_be.repository.PlanRepository;
+import com.SOLUX_WEBFINITE_BE.webfinite_be.repository.PromptRepository;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonMappingException;
